requirements.txt
flask
flask-cors
easyocr
numpy
torch
python-levenshtein
Pillow